Simplified splitter.
